+++ sdk/healthdataaiservices/azure-health-deidentification/azure/health/deidentification/_operations/_operations.py
@@ -518,7 +518,7 @@ class DeidentificationClientOperationsMixin(DeidentificationClientMixinABC):
     @distributed_trace
     def _list_jobs_internal(
         self, *, continuation_token_parameter: Optional[str] = None, **kwargs: Any
-    ) -> Iterable["_models.DeidentificationJob"]:
+    ) -> ItemPaged["_models.DeidentificationJob"]:
         """List de-identification jobs.
 
         Resource list operation template.
@@ -610,7 +610,7 @@ class DeidentificationClientOperationsMixin(DeidentificationClientMixinABC):
     @distributed_trace
     def _list_job_documents_internal(
         self, job_name: str, *, continuation_token_parameter: Optional[str] = None, **kwargs: Any
-    ) -> Iterable["_models.DeidentificationDocumentDetails"]:
+    ) -> ItemPaged["_models.DeidentificationDocumentDetails"]:
         """List processed documents within a job.
 
         Resource list operation template.